Fix layer breaker
